Optimizing and features
Easy to hit FB API limit when calling entire inbox each time. Now only retrieving threads with activity in the past few minutes.

Iterating the number of actual messages get sent from FB to Google Talk and tally up the numbers at the end and use that number as the last line of the method along with that being the only output.