Cookie information stored in db versus file
Heroku doesn't allow non-temporary reading or writing of files. So cookie information moved from being kept in a file to db.

A cookie corresponds to a user so info stored in a column for a specific user (FbMember table).

Also added Facebook username and Facebook password columns. Code looks for a specific Facebook user when trying to get cookie information.